Fix restart of O32 indirect syscalls
When an indirect syscall is restarted a0..a3 contain the target
syscall arguments but any additional arguments remain on the user
stack at the original offset. Saving the target system call as a
negative value in v0 allows bot the target system call to be recovered
and as a flag to locate the remaining stack arguments.

Change-Id: I5bf3f5c0d2a65847399f86e492c3ff7675d10e0a
Signed-off-by: Chris Dearman <chris.dearman@imgtec.com>