Port some shrinker-related fixes from common/android-3.4.
This is a manual rediff of some changes made upstream in android-3.4
to factor the shrinker out of the vm/fs code so that it can be used
by drivers.

The shrinker is needed by ion which will be added in a subsequent
commit.

I don't recommend carrying this change permanently. It would certainly
be better if android-3.4 was imported into NPM801, but that is not
easy for me to do because the NPM801 kernel descends from a weird
non-Linus parent commit, and so a trivial "git merge" is not possible.