Add support for the Ingenic 'xburst' platform to ion.
This is just an example of how to integrate ion. Normally the ion
heap definition would be done in the arch board code, but as Ingenic's
requirements are very simple (one heap only, no carved out memory) I
decided against it.

The MIPS portability problems this diff corrects in ion have been fixed
in a more "mainline" way in android-3.10, but they don't back-port
cleanly to 3.4.

Tested with the PowerVR SGX driver.