Update NAND Driver to support NAND-Manager 1.6 for the UR-BOARD.
The NPM801 uses NAND-MANAGER (NM) 1.4.2. The 1st UR-BOARD,
without a JTAG Header, used LM-V1.6.1 and the new UR-BOARD
with a JTAG Header uses LM-V1.6.2. The NM version is dependent
on the version of xboot being used.

I updated the complete nand driver to that containing the
NM-1.6.2 code as it appears to include only generic improvements.
The old 1.4.2 LM code was renamed from 'manager' to 'manager_v1.4.2'
for clarity at the expense of a larger git commit.

The new NM-v1.6.2 was added to a new 'manager_1.6.2' directory.
I suspect that this 1.6.2 manager code is likely compatible with
various 1.6 versions of the NM library and a symlink was added:

     manager_v1.6.1 --> manager_v1.6.2

Change-Id: I7eea6aacc83074664357aa95739d92c5e7413597
Signed-off-by: Pete Delaney <piet.delaney@imgtec.com>